Reduce application memory footprint
Why:
Each time a pick was made or undone, a call was made to the drafts_controller to fetch the new draft
state (picks, players, etc.). This turned out to be a significant amount of data, particularly for
baseball, which has roughly five times as many players as football.

This design will not allow for a reactive application running on basic hardware, as the memory consumed
by such a load for 12 parallel users is too high.

How:
Manage the draft state on the client. After the initial draft room load, responses from the server
should only pass the incremental data required to maintain said state.