Upgrade to Rails 5.
Among the changes required when upgrading from the beta to the official release, was
a modification to the react debounce on the player search input. The new approach is
derived from the answer to: http://stackoverflow.com/questions/23123138/perform-debounce-in-react-js

There were substantial issues encountered with server-side redirects. These either caused a
JSON parse error, because the AJAX requests were expecting a JSON response, but were
receiving a script (the Turbolinks redirect), or did a full page reload, thus negating the
performance advantages of using React. The revised approach is an attempt to maintain the
advantages of using the Virtual DOM, while minimizing repetition of server-side code.