Remove inactive players from the player pool.
It should be noted that I attempted to do this a year ago, but ran into an issue where the CBS API
removes players who are on temporary bereavement from the player pool. While I can't recall exactly
how I figured that out, there's currently no mention of the bereavement list in the API docs. Even
in the event that the issue still exists, it's likely worthwile to make this change for the performance
gains.